requirements.txt
flask
flask-cors
gunicorn
cohere
python-dotenv
